ignored ink json
Added main ink file to gitignore. Since this file only represents the compiled version of all ink's scripts, it is always recreated locally, and always changes with each ink script, meaning a lot of ultimately needless changes get commited on this file which had to be stashed (or needlessly clutter up the changelog). small annoyances build up, and so adding this to gitignore will hopefully avoid some annoyance in future.