Continued work on Coastal Ruins
Currently building the first dungeon (in draft pieces, will move that later). Started with a corridor in the middle of the dungeon because that's how my brain works, apparently. Done most of the work for the navigating of the corridor, will start adding the actual encounter in there in my next session.
@BassWasTaken I left a note for you in there because I couldn't work out a thingy with the variables. Would be great if you can take a look!