made choice buttons focus compatible
issue was: choice buttons were not receiving the activation, because space was instead bneing read as continue
solution: i don't fully know tbh but i think it's worked because i dfecoupled some of the continue firing, and it now fires just when the button is presed. and space pressed that button. yeah you know now that i say this, why did i ever do it differently? that's 2 ways, this is just 1